+++ src/EntwineLLM/Commands/BaseCommand.cs
@@ -1,8 +1,10 @@
 ﻿using EntwineLlm.Enums;
 using EntwineLlm.Helpers;
+using ICSharpCode.AvalonEdit;
 using Microsoft.VisualStudio.Shell;
 using System;
 using System.Threading.Tasks;
+using System.Windows.Controls;
 
 namespace EntwineLlm.Commands
 {
@@ -11,6 +13,8 @@ namespace EntwineLlm.Commands
         public AsyncPackage package;
 
         public string ActiveDocumentPath;
+        public TextBox ManualPromptTextBox;
+        public TextEditor SuggestedCodeEditor;
 
         public BaseCommand(AsyncPackage package)
         {
@@ -41,16 +45,19 @@ namespace EntwineLlm.Commands
             return textSelection.Text;
         }
 
-        public async Task PerformRefactoringSuggestionAsync(RequestedCodeType codeType)
+        public async Task PerformRefactoringSuggestionAsync(RequestedCodeType codeType, string manualPrompt = "")
         {
             var message = "Waiting for LLM response (task requested: " + Enum.GetName(typeof(RequestedCodeType), codeType) + ") ...";
-            
+
             var progressBarHelper = new ProgressBarHelper(ServiceProvider.GlobalProvider);
             progressBarHelper.StartIndeterminateDialog(message);
 
-            var methodCode = GetCurrentMethodCode();
+            var methodCode = SuggestedCodeEditor != null ?
+                SuggestedCodeEditor.Text
+                : GetCurrentMethodCode();
+
             var refactoringHelper = new RefactoringHelper(package);
-            await refactoringHelper.RequestCodeSuggestionsAsync(methodCode, ActiveDocumentPath, codeType);
+            await refactoringHelper.RequestCodeSuggestionsAsync(methodCode, ActiveDocumentPath, codeType, manualPrompt);
 
             progressBarHelper.StopDialog();
         }