+++ src/EntwineLLM/EntwineLLMPackage.cs
@@ -18,9 +18,12 @@ namespace EntwineLlm
     {
         public const string PackageGuidString = "3c995b0e-1f37-4cef-9ac7-9771b3fb6162";
 
+        public static AsyncPackage Instance { get; set; }
+
         protected override async Task InitializeAsync(CancellationToken cancellationToken, IProgress<ServiceProgressData> progress)
         {
             await JoinableTaskFactory.SwitchToMainThreadAsync(cancellationToken);
+            Instance = this;
 
             var commandsMenu = new CommandsMenu();
             await commandsMenu.InitializeAsync(this);