+++ src/EntwineLLM/Helpers/PromptHelper.cs
@@ -1,9 +1,17 @@
 ﻿using EntwineLlm.Extensions;
+using EntwineLlm.Models;
+using Newtonsoft.Json;
 
 namespace EntwineLlm.Helpers
 {
     internal static class PromptHelper
     {
+        public static string CreateForManualRequest(string model, string userCode, string prompt)
+        {
+            var promptText = PreparePrompt(Properties.Resources.PromptForManual);
+            return ReplacePlaceholders(promptText, model, userCode, prompt);
+        }
+
         public static string CreateForRefactor(string model, string userCode)
         {
             var promptText = PreparePrompt(Properties.Resources.PromptForRefactor);
@@ -23,12 +31,17 @@ namespace EntwineLlm.Helpers
                 .ReduceMultipleSpaces();
         }
 
-        private static string ReplacePlaceholders(string prompt, string model, string code)
+        private static string ReplacePlaceholders(string prompt, string model, string code, string manualRequest = "")
         {
-            return Properties.Resources.LlmBaseRequest
-                    .Replace("{PROMPT}", prompt)
-                    .Replace("{MODEL}", model)
-                    .Replace("{CODE}", code.EscapeJsonString());
+            code = code.EscapeJsonString();
+            manualRequest = manualRequest.EscapeJsonString();
+
+            var llmRequest = LlmRequest.Create(model, prompt, code, manualRequest);
+            return JsonConvert.SerializeObject(llmRequest, Formatting.Indented, new JsonSerializerSettings()
+            {
+                MissingMemberHandling = MissingMemberHandling.Ignore,
+                NullValueHandling = NullValueHandling.Ignore
+            });
         }
     }
 }